+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-09T07:05:28.338Z",
+    "2026-02-09T10:59:48.228Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {