+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-09T07:05:25.001Z",
+    "2026-02-09T10:59:44.531Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {