+++ packages/orq-rc/RELEASES.md
@@ -6529,3 +6529,13 @@ Based on:
 - [typescript v4.3.0-rc.38] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.38 - packages/orq-rc
+
+## 2026-02-09 11:02:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.1 (2.812.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.39] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.39 - packages/orq-rc